requirements.txt
lxml-html-clean==0.4.3
beautifulsoup4==4.14.3
fake-useragent==2.2.0
newspaper4k==0.9.4.1
tenacity==9.1.2
pandas==2.3.3
numpy==2.4.0
nltk==3.9.2
tqdm==4.67.1
aiohttp>=3.11.0
aiofiles>=24.1.0
pyarrow>=19.0.0
joblib>=1.4.0